sql-passes: Move autoparametrization pass to its own module
This commit moves the autoparametrization rewrite pass to its own
module. It already has quite a bit of code, and it will only grow larger
once we add support for the autoparametrization of range queries in a
future commit.
